fix i18n function
